divide charge_mixing.cpp into some files
